Fix order of assigning coordinates and velocities for different types of stars
