+++ alcor/services/simulations/velocities.py
@@ -4,6 +4,7 @@ from typing import Tuple
 
 import numpy as np
 import pandas as pd
+import pydevd
 
 from alcor.models.star import GalacticDiskType
 from alcor.types import GaussianGeneratorType
@@ -24,12 +25,11 @@ def set_velocities(stars: pd.DataFrame,
                    oort_constant_b: float,
                    generator: GaussianGeneratorType = np.random.normal
                    ) -> None:
+    # pydevd.settrace('dockerhost', port=20111)
     thin_disk_stars_mask = (stars['galactic_disk_type']
                             == GalacticDiskType.thin)
     thick_disk_stars_mask = (stars['galactic_disk_type']
                              == GalacticDiskType.thick)
-
-    # (white_dwarfs.loc[halo_stars_mask, 'u_velocity']
 
     stars_velocities = partial(
             disk_stars_velocities,
